requirements.txt
psycopg2==2.9.6
lxml==4.9.3
requests==2.31.0
nmap==1.5.0
python-nmap==0.7.1
beautifulsoup4==4.12.2
PyYAML==6.0
